flight_tracker/requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
pyserial>=3.5
openpyxl>=3.1.0
python-engineio>=4.5.0
python-socketio>=5.8.0
